refactor: resolver protector 패키지 이동
