+++ src/main/java/com/sign/controller/argumentresolver/EmailTokenResolver.java
@@ -2,7 +2,6 @@ package com.sign.controller.argumentresolver;
 
 import com.sign.controller.CookieManager;
 import com.sign.controller.ProtectedRequest;
-import com.sign.controller.ResponseProtector;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
