+++ src/main/java/com/sign/controller/config/WebMvcConfig.java
@@ -1,10 +1,10 @@
 package com.sign.controller.config;
 
 import com.sign.controller.CookieManager;
-import com.sign.controller.ResponseProtector;
 import com.sign.controller.argumentresolver.EmailTokenResolver;
 import com.sign.controller.argumentresolver.PassKeyRegistrationResolver;
 import com.sign.controller.argumentresolver.PasskeyAssertionResolver;
+import com.sign.controller.argumentresolver.ResponseProtector;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
